Add debug logs and refactor API requests
Added debug logging to getCurrentUserCredentials query and AuthServiceDataSource methods. This was done to improve the visibility of what's happening inside these modules for easier debugging and tracing of potential issues during token validation or password comparison.

Additionally, the way we send request parameters to AuthService's API endpoints has been refactored. Instead of sending the parameters in the request body, we are now sending them as query parameters, hence the change from `this.get` with `body` to `this.get` with `params`. This change has been made to comply with recent updates in the AuthService API specification, where all GET requests now must include their parameters in the URL query string.