Replace request body with query parameters and add debug logging
The commit contains changes made in the AuthServiceDataSource file to add debug logging and refactor the API requests. Debug logging has been introduced to track the flow of processes in the AuthServiceDataSource methods. This will enhance error tracing in token validation and password comparison. The sending of request parameters has been refactored from being part of the request body to being conveyed as query parameters. This change is to adhere to the updated specification of the AuthService API, where all GET requests must have their parameters in the URL query string.