Remove package-lock.json file
The package-lock.json file has been deleted from the chatService directory as it should not be version controlled. This file is automatically generated for any operations where npm modifies either the node_modules tree, or package.json, hence doesn't need to be tracked. This will also prevent any merge conflicts in the future related to this file.