Refactored code to set the default status of contact request as 'pending'
The code was modified to set the status of a new contact request to 'pending' as a default. To achieve this, the 'status' parameter was removed where we were creating a new contact request object, in the sendContactRequest functions in ContactController.js and ContactServiceDataSource.js. The status was directly set as 'pending' in the ContactRequest object in ContactController.js. This change eliminates the need to manually specify 'pending' as in status while making a request. The acknowledgment here is to make the process of creating new contact requests less error-prone, more streamlined, and intuitive.