Added accept contact request functionality
Multiple changes have been made throughout the code to add the functionality of accepting a contact request. This includes updates in the GraphQL type definitions and resolvers, adding a new 'acceptContactRequest' function and relevant methods in the AuthServiceDataSource.js, ContactServiceDataSource.js and dataSources files. Updates also include changes in the UserController.js to handle the process of a user accepting a contact request and a new field 'contacts' has been added to the UserModel.js. Moreover, necessary changes have been made in the ChatController.js and ContactController.js to handle the new functionality. This functionality was essential to handle the user interaction of accepting a contact request, storing the request and updating the user's contacts list accordingly.