Remove package-lock.json file
Deleting the package-lock.json file as it is not needed. This file is generated and updated automatically whenever we run npm install. It locks the installed dependencies to a specific version, and ensures that future installs are able to reproduce the same tree structure. However, it can cause merge conflicts and we can safely remove it from the repository.