Add functionality to retrieve chatrooms for logged in user
Added the ability for a logged user to retrieve a list of associated chat rooms. A new GraphQL query getChatRoomsForCurrentUser was created for this purpose. Also, a new route was added in the Chat service for finding chat rooms by user id. This improvement will streamline the user's experience by readily providing all relevant chat rooms.