Add getContactRequestsByContext function to fetch all contact requests
An update was necessary to fetch all contact requests by context for improving user experience and functionality. A new type 'ContactRequestPopulated' was defined in the graphql/typeDefs.js file. The function was then added to the resolvers.js file for execution. getContactRequestByContext function was renamed to getContactRequestsByContext in queries for syntactic correctness and increased usability. Debug elements were added in a number of places to aid in tracking function execution and identifying potential errors. Changes were also reflected in the frontend gql query for compatibility. The ContactAPI service was adjusted for better syntax alignment.