Add feature to fetch contacts with chat room
This commit adds a feature that enables fetching of contacts in conjunction with chat rooms. A new type 'UserWithChatRoom' was defined in the graphql/typeDefs.js file, along with a new query 'getContactsWithChatRoom'. A new file 'getContactsWithChatRoom.js' was added with the logic for fetching contacts and its chat room. The resolver.js file was updated to include this new query. Endpoint '/v1/contacts' was added to auth service which is used to fetch contacts of a user. The relevant functions were also updated or created in 'UserController.js' and 'AuthServiceDataSource.js'. This functionality was needed for a smoother user experience and better data handling.