Change base URLs to use environment variables
Instead of hardcoding the base URLs in the dataSources classes (AuthServiceDataSource.js, ChatServiceDataSource.js and ContactServiceDataSource.js), they are now fetched from the environment variables if available. This allows for flexibility in development, testing, and deployment as the base URLs can easily be changed based on the environment. Docker-compose.yaml file is also updated to include these environment variables.