Refactor password and token handlers, improve code readability
This commit mainly refactors the handlers for getting tokens and passwords in the UserController and PasswordController. Previously, the functionalities of comparing, encrypting, and getting password were separated into different endpoints, making the code more complex and less user-friendly. This commit merges these functionalities into a single endpoint for both password and token, which are now based on the query parameters. Changed methods names are also more precise and reflective of their behaviors. As a result, the code becomes cleaner and easier to maintain.

This commit also updates the corresponding GraphQL queries in AuthServiceDataSource. The paths of these queries are modified to match the changed endpoints.

The old endpoints and the associated handlers are commented out rather than deleted. They can be safely removed in the future if the new design has been proved to be stable and efficient.