Add auth-service deployment and service configuration
This commit introduces two new configuration files for Kubernetes deployment of the 'auth-service'. The `auth_service-service.yaml` file details specifications for the service, including the TCP protocol and port to be used. The `auth_service-deployment.yaml` file handles the deployment of the service, including the Docker image to be used and the environment variables necessary for the service's operation. Minor indentation adjustment is also done on the `mongo_auth-deployment.yaml` file.