Add password validation in user creation
The new file 'passwordValidator.js' is added for implementing password validation during user creation. The 'createUser' function in 'UserController.js' is updated to utilize the validator and send a response if the password does not meet the required criteria. It improves the security of user creation by enforcing password strength rules.