Add Signup component tests
A new file named Signup.test.js was added to provide unit tests for the Signup component. Tests include checking the proper rendering of the signup form and the ability to update the state when the form fields are changed. More similar tests will be added to cover different aspects of the component behavior.